main rebuilt marketlink_core (responses, exceptions, http, pagination,
permissions, policies, services/db_retry) and switched the currency to USD.
Resolved in main's favour for the platform, then re-pointed the admin and
public code at it.

Core layer
- marketlink_core.utils dropped; api_response and custom_exception_handler
  now come from marketlink_core.responses. audit_request moved to
  manager.common.audit, over system.services.log_request_event.
- manager.common.retry dropped for marketlink_core.services.db_retry.
- ContractPagination dropped for StandardPagination; the public review lists
  use PublicReviewPagination (10 per page, Pass 4B §2.2).
- Error catalog follows Pass 4B §2.5: 401 NOT_AUTHENTICATED, 403
  PERMISSION_DENIED, 409 RESOURCE_MODIFIED, 429 THROTTLED; a duplicate row
  reads as 400 VALIDATION_ERROR and a foreign key clash as 422
  RESOURCE_IN_USE.
- system.services.log_security_event validates ip_address again: the column
  only fits an IP, so a caller passing a raw header must not cause a 500.

Currency, D-020 / Change Log v1.2
- Money leaves the API as a 2-decimal string ("12.50"), not a number:
  ProductCard.price, OrderSummary.total_amount and the report revenues.
- The report export writes numeric USD cells under "Revenue (USD)".
- The rows that block an action ride in errors, per FA-32 (errors.order_ids):
  AD-17 deactivate (D-022) and AD-32 closures (D-023).
- frontend formatCurrency now renders $1,245.50.
- Test fixtures moved from VND to USD amounts, inside the new
  prod_price_range check ($0.01-$10,000.00).

CLAUDE.md keeps both sets of rules; the skill moved to Document/skill/SKILL.md.

Checked against: D-020, D-022, D-023, Pass 4B §2.1, §2.2, §2.4, §2.5, §3.4,
AD-17, AD-32, FA-32. 241 tests pass on MySQL 8, ruff clean, migrations apply.

+++ backend/catalog/public/serializers_public.py
@@ -18,7 +18,7 @@ class CategoryReadSerializer(serializers.ModelSerializer):
 class ProductCardSerializer(serializers.ModelSerializer):
     """ProductCard (Pass 4B §3.3). Needs catalog.public.products.public_products()."""
 
-    price = serializers.IntegerField()
+    price = serializers.DecimalField(max_digits=10, decimal_places=2)
     availability = serializers.SerializerMethodField()
     category = serializers.SerializerMethodField()
     farmer = serializers.SerializerMethodField()
